rearranged file, and renamed files and components
